requirements.txt
# Authoricy SEO Analyzer - Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Async HTTP client
httpx>=0.25.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Claude AI
anthropic>=0.7.0

# PDF generation
weasyprint>=60.0

# Email
resend>=0.7.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging
structlog>=23.2.0
